+++ src/components/app/app.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import MainScreen from '../pages/main-screen/main-screen';
-import { AppRoute, Status } from '../constants';
-import FavoritesScreen from '../pages/favorites/favorites-screen';
-import LoginScreen from '../pages/login-screen/login-screen';
-import NotFoundScreen from '../pages/not-found-screen/not-found-screen';
-import OfferScreen from '../pages/offer/offer-screen';
+import MainScreen from '../../pages/main-screen/main-screen';
+import { AppRoute, Status } from '../utils/constants';
+import FavoritesScreen from '../../pages/favorites/favorites-screen';
+import LoginScreen from '../../pages/login-screen/login-screen';
+import NotFoundScreen from '../../pages/not-found-screen/not-found-screen';
+import OfferScreen from '../../pages/offer/offer-screen';
 import PrivateRoute from '../private-route/private-route';
 import { Offer } from '../../types/offer';
 import { Review } from '../../types/review';
