+++ src/components/private-route/private-route.tsx
@@ -1,6 +1,6 @@
 import {Navigate} from 'react-router-dom';
-import { Status } from '../constants';
-import { AppRoute } from '../constants';
+import { Status } from '../utils/constants';
+import { AppRoute } from '../utils/constants';
 
 type PrivateRouteProps = {
   status: Status;
