+++ src/pages/favorites/favorites-empty-screen.tsx
