+++ src/pages/favorites/favorites-screen.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
-import { Offer } from '../../../types/offer';
-import Card from '../../card/card';
+import { Offer } from '../../types/offer';
+import Card from '../../components/card/card';
 
 
 type FavoritesScreenProps = {
